Create makemicrowave.plx
This file spits out the text needed to make a long run of arcs, in long form. I couldn't get gnuplot's "do for" loop to behave in a way that made the arcs on my output plots. Thus, this file makes a big block of text that one can copy and paste into the gnuplot script.